Moved codemirror into the existing src folder
