--- codemirror-lang-sequence/.gitignore
@@ -1,4 +1,0 @@
-.DS_Store
-/node_modules
-/dist
-/src/*.js
