--- codemirror-lang-sequence/.npmignore
@@ -1,5 +1,0 @@
-/src
-/test
-/node_modules
-rollup.config.js
-tsconfig.json
