--- codemirror-lang-sequence/.prettierrc
@@ -1,7 +1,0 @@
-{
-  "printWidth": 120,
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "all",
-  "useTabs": false
-}
