--- codemirror-lang-sequence/README.md
@@ -1,1 +1,0 @@
-# codemirror-lang-seq
