--- codemirror-lang-sequence/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "codemirror-lang-sequence",
-  "version": "0.0.1",
-  "type": "module",
-  "license": "MIT",
-  "types": "dist/index.d.ts",
-  "main": "dist/index.cjs",
-  "module": "dist/index.js",
-  "sideEffects": false,
-  "exports": {
-    "import": "./dist/index.js",
-    "require": "./dist/index.cjs"
-  },
-  "scripts": {
-    "build": "npm run clean && rollup -c",
-    "clean": "rm -rf dist",
-    "postinstall": "[ \"$NODE_ENV\" = production ] && exit 0; npm run build",
-    "test": "npm run build && mocha test/*.js"
-  },
-  "dependencies": {
-    "@codemirror/autocomplete": "^6.15.0",
-    "@codemirror/language": "^6.8.0",
-    "@lezer/highlight": "^1.1.6",
-    "@lezer/lr": "^1.3.9"
-  },
-  "devDependencies": {
-    "@lezer/generator": "^1.3.0",
-    "mocha": "^10.2.0",
-    "prettier": "^3.0.0",
-    "rollup": "^3.26.3",
-    "rollup-plugin-dts": "^5.3.0",
-    "rollup-plugin-ts": "^3.2.0",
-    "typescript": "^5.1.6"
-  },
-  "description": "",
-  "directories": {
-    "test": "test"
-  },
-  "author": ""
-}
