--- codemirror-lang-sequence/test/cases/errors.txt
@@ -1,49 +1,0 @@
-# Bad Input - Invalid stems
-
-C 2_STEM_NAME
-STEM$BAR
-
-==>
-
-Sequence(Commands(Command(TimeTag(TimeComplete),⚠(Number),Stem,Args),Command(Stem,⚠),Command(Stem,Args)))
-
-# Stem with disallowed characters
-
-FSW_CMD%BAR$BAZ
-
-==>
-Sequence(Commands(
-  Command(Stem,⚠),
-  Command(Stem,⚠),
-  Command(Stem,Args)
-))
-
-# Stem ending in disallowed character
-
-FSW_CMD%
-
-==>
-Sequence(Commands(
-  Command(Stem,⚠,Args)
-))
-
-# Mismatched brackets
-
-CMD [[]
-CMD2 [
-CMD3 ]
-
-==>
-
-Sequence(Commands(
-  Command(Stem,Args(RepeatArg(⚠))),
-  Command(Stem,Args(RepeatArg(⚠,Enum)))
-))
-
-# locals with wrong value types
-
-@LOCALS "string_not_enum"
-
-==>
-
-Sequence(LocalDeclaration(⚠(String)))
