--- codemirror-lang-sequence/test/sequences/seq_with_empty_lines.txt
@@ -1,13 +1,0 @@
-
-# standalone comment
-CMD_1 "strarg" 0 # inline comment
-
-# standalone comment
-
-
-
-
-CMD_2 "strarg" 0 # inline comment
-
-@METADATA "key" "val"
-
