--- codemirror-lang-sequence/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "module": "es2020",
-    "moduleResolution": "node",
-    "newLine": "lf",
-    "strict": true,
-    "target": "es6"
-  },
-  "include": ["src/*.ts"]
-}
