+++ package.json
@@ -77,9 +77,13 @@
     "toastify-js": "^1.12.0"
   },
   "devDependencies": {
-    "@codemirror/autocomplete": "^6.15.0",
+    "@codemirror/autocomplete": "^6.16.0",
     "@codemirror/lang-json": "^6.0.1",
+    "@codemirror/language": "^6.10.1",
     "@codemirror/lint": "^6.5.0",
+    "@lezer/generator": "^1.7.0",
+    "@lezer/highlight": "^1.2.0",
+    "@lezer/lr": "^1.4.0",
     "@playwright/test": "^1.42.1",
     "@poppanator/sveltekit-svg": "^4.2.1",
     "@sveltejs/vite-plugin-svelte": "^3.0.0",
@@ -108,7 +112,6 @@
     "@vitest/ui": "^1.4.0",
     "cloc": "2.0.0-cloc",
     "codemirror": "^6.0.1",
-    "codemirror-lang-sequence": "./codemirror-lang-sequence",
     "d3-format": "^3.1.0",
     "d3-zoom": "^3.0.0",
     "esbuild": "^0.20.2",
@@ -116,6 +119,7 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-svelte": "^2.32.0",
     "jsdom": "^24.0.0",
+    "mocha": "^10.4.0",
     "picocolors": "^1.0.0",
     "picomatch": "^4.0.1",
     "postcss-html": "^1.5.0",
@@ -131,8 +135,5 @@
     "unique-names-generator": "^4.7.1",
     "vite": "5.1.6",
     "vitest": "^1.0.0"
-  },
-  "workspaces": [
-    "codemirror-lang-sequence"
-  ]
+  }
 }
