+++ src/utilities/codemirror/sequence.grammar
