+++ src/utilities/codemirror/sequence.grammar.d.ts
